+++ packages/context-selector/src/index.tsx
@@ -1,3 +1,5 @@
+export * from './ContextProvider';
 export * from './ContextSelector';
+export * from './types';
 
 export { ContextSelector as default } from './ContextSelector';