+++ packages/searchable-dropdown/src/Dropdown.tsx
@@ -1,7 +1,12 @@
 /* eslint-disable react/no-multi-comp */
+import { BaseSyntheticEvent } from 'react';
 import { ComponentProps, createComponent } from '@equinor/fusion-react-utils';
 
-import { SearchableDropdownElement } from '@equinor/fusion-wc-searchable-dropdown';
+import {
+  SearchableDropdownElement,
+  SearchableDropdownSelectEvent as HTMLSearchableDropdownSelectEvent,
+} from '@equinor/fusion-wc-searchable-dropdown';
+
 export {
   SearchableDropdownResult,
   SearchableDropdownResultItem,
@@ -11,12 +16,24 @@ export {
 type ElementAttributes = Partial<
   Pick<
     SearchableDropdownElement,
-    'label' | 'placeholder' | 'variant' | 'meta' | 'selected' | 'initialText' | 'trailingIcon'
+    | 'id'
+    | 'label'
+    | 'placeholder'
+    | 'variant'
+    | 'meta'
+    | 'value'
+    | 'initialText'
+    | 'leadingIcon'
+    | 'dropdownHeight'
+    | 'multiple'
+    | 'graphic'
   >
 >;
 
+export type SearchableDropdownSelectEvent = BaseSyntheticEvent<HTMLSearchableDropdownSelectEvent>;
+
 type ElementEvents = {
-  onSelect?: (e: Event) => void;
+  onSelect?: (e: SearchableDropdownSelectEvent) => void;
 };
 
 type ElementProps = ElementAttributes & ElementEvents;